requirements.txt
# Core dependencies (aligned with pyproject.toml)
cryptography>=44.0.1
fastapi>=0.115.8
httpx>=0.28.1
matplotlib>=3.10.0
numpy>=2.2.2
openai>=1.61.0
pandas>=2.2.3
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.7.1
pyjwt>=2.10.1
pyopenssl>=25.0.0
requests>=2.32.3

structlog>=25.1.0
uvicorn>=0.34.0

# Testing and development dependencies
scikit-learn>=1.3.0
tqdm>=4.66.0

# Telegram Bot
python-telegram-bot>=20.8
python-dotenv==1.0.0

# RAG System
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
pinecone-client>=2.2.4
tiktoken>=0.5.2